requirements.txt
boto3
botocore
odc-geo
odc-stac
pystac_client
typer

